@@ -74,7 +74,7 @@ class Collator
      *
      * @throws MethodArgumentValueNotImplementedException When $locale different than "en" or null is passed
      */
-    public function __construct($locale)
+    public function __construct(?string $locale)
     {
         if ('en' !== $locale && null !== $locale) {
             throw new MethodArgumentValueNotImplementedException(__METHOD__, 'locale', $locale, 'Only the locale "en" is supported');
@@ -108,11 +108,11 @@ class Collator
      */
     public function asort(&$array, $sortFlag = self::SORT_REGULAR)
     {
-        $intlToPlainFlagMap = array(
+        $intlToPlainFlagMap = [
             self::SORT_REGULAR => \SORT_REGULAR,
             self::SORT_NUMERIC => \SORT_NUMERIC,
             self::SORT_STRING => \SORT_STRING,
-        );
+        ];
 
         $plainSortFlag = isset($intlToPlainFlagMap[$sortFlag]) ? $intlToPlainFlagMap[$sortFlag] : self::SORT_REGULAR;
 
