@@ -11,8 +11,8 @@
 
 namespace Symfony\Component\Intl\Data\Generator;
 
-use Symfony\Component\Intl\Data\Bundle\Compiler\GenrbCompiler;
-use Symfony\Component\Intl\Data\Bundle\Reader\BundleReaderInterface;
+use Symfony\Component\Intl\Data\Bundle\Compiler\BundleCompilerInterface;
+use Symfony\Component\Intl\Data\Bundle\Reader\BundleEntryReaderInterface;
 use Symfony\Component\Intl\Data\Util\ArrayAccessibleResourceBundle;
 use Symfony\Component\Intl\Data\Util\LocaleScanner;
 
@@ -25,45 +25,28 @@ use Symfony\Component\Intl\Data\Util\LocaleScanner;
  */
 class CurrencyDataGenerator extends AbstractDataGenerator
 {
-    const UNKNOWN_CURRENCY_ID = 'XXX';
-    const EUROPEAN_COMPOSITE_UNIT_ID = 'XBA';
-    const EUROPEAN_MONETARY_UNIT_ID = 'XBB';
-    const EUROPEAN_UNIT_OF_ACCOUNT_XBC_ID = 'XBC';
-    const EUROPEAN_UNIT_OF_ACCOUNT_XBD_ID = 'XBD';
-    const TESTING_CURRENCY_CODE_ID = 'XTS';
-    const ADB_UNIT_OF_ACCOUNT_ID = 'XUA';
-    const GOLD_ID = 'XAU';
-    const SILVER_ID = 'XAG';
-    const PLATINUM_ID = 'XPT';
-    const PALLADIUM_ID = 'XPD';
-    const SUCRE_ID = 'XSU';
-    const SPECIAL_DRAWING_RIGHTS_ID = 'XDR';
-
-    /**
-     * Monetary units excluded from generation.
-     */
-    private static $blacklist = array(
-        self::UNKNOWN_CURRENCY_ID => true,
-        self::EUROPEAN_COMPOSITE_UNIT_ID => true,
-        self::EUROPEAN_MONETARY_UNIT_ID => true,
-        self::EUROPEAN_UNIT_OF_ACCOUNT_XBC_ID => true,
-        self::EUROPEAN_UNIT_OF_ACCOUNT_XBD_ID => true,
-        self::TESTING_CURRENCY_CODE_ID => true,
-        self::ADB_UNIT_OF_ACCOUNT_ID => true,
-        self::GOLD_ID => true,
-        self::SILVER_ID => true,
-        self::PLATINUM_ID => true,
-        self::PALLADIUM_ID => true,
-        self::SUCRE_ID => true,
-        self::SPECIAL_DRAWING_RIGHTS_ID => true,
-    );
+    private static $blacklist = [
+        'XBA' => true, // European Composite Unit
+        'XBB' => true, // European Monetary Unit
+        'XBC' => true, // European Unit of Account (XBC)
+        'XBD' => true, // European Unit of Account (XBD)
+        'XUA' => true, // ADB Unit of Account
+        'XAU' => true, // Gold
+        'XAG' => true, // Silver
+        'XPT' => true, // Platinum
+        'XPD' => true, // Palladium
+        'XSU' => true, // Sucre
+        'XDR' => true, // Special Drawing Rights
+        'XTS' => true, // Testing Currency Code
+        'XXX' => true, // Unknown Currency
+    ];
 
     /**
      * Collects all available currency codes.
      *
      * @var string[]
      */
-    private $currencyCodes = array();
+    private $currencyCodes = [];
 
     /**
      * {@inheritdoc}
@@ -76,7 +59,7 @@ class CurrencyDataGenerator extends AbstractDataGenerator
     /**
      * {@inheritdoc}
      */
-    protected function compileTemporaryBundles(GenrbCompiler $compiler, $sourceDir, $tempDir)
+    protected function compileTemporaryBundles(BundleCompilerInterface $compiler, $sourceDir, $tempDir)
     {
         $compiler->compile($sourceDir.'/curr', $tempDir);
         $compiler->compile($sourceDir.'/misc/currencyNumericCodes.txt', $tempDir);
@@ -87,21 +70,21 @@ class CurrencyDataGenerator extends AbstractDataGenerator
      */
     protected function preGenerate()
     {
-        $this->currencyCodes = array();
+        $this->currencyCodes = [];
     }
 
     /**
      * {@inheritdoc}
      */
-    protected function generateDataForLocale(BundleReaderInterface $reader, $tempDir, $displayLocale)
+    protected function generateDataForLocale(BundleEntryReaderInterface $reader, $tempDir, $displayLocale)
     {
         $localeBundle = $reader->read($tempDir, $displayLocale);
 
         if (isset($localeBundle['Currencies']) && null !== $localeBundle['Currencies']) {
-            $data = array(
+            $data = [
                 'Version' => $localeBundle['Version'],
                 'Names' => $this->generateSymbolNamePairs($localeBundle),
-            );
+            ];
 
             $this->currencyCodes = array_merge($this->currencyCodes, array_keys($data['Names']));
 
@@ -112,20 +95,20 @@ class CurrencyDataGenerator extends AbstractDataGenerator
     /**
      * {@inheritdoc}
      */
-    protected function generateDataForRoot(BundleReaderInterface $reader, $tempDir)
+    protected function generateDataForRoot(BundleEntryReaderInterface $reader, $tempDir)
     {
         $rootBundle = $reader->read($tempDir, 'root');
 
-        return array(
+        return [
             'Version' => $rootBundle['Version'],
             'Names' => $this->generateSymbolNamePairs($rootBundle),
-        );
+        ];
     }
 
     /**
      * {@inheritdoc}
      */
-    protected function generateDataForMeta(BundleReaderInterface $reader, $tempDir)
+    protected function generateDataForMeta(BundleEntryReaderInterface $reader, $tempDir)
     {
         $rootBundle = $reader->read($tempDir, 'root');
         $supplementalDataBundle = $reader->read($tempDir, 'supplementalData');
@@ -135,12 +118,12 @@ class CurrencyDataGenerator extends AbstractDataGenerator
 
         sort($this->currencyCodes);
 
-        $data = array(
+        $data = [
             'Version' => $rootBundle['Version'],
             'Currencies' => $this->currencyCodes,
             'Meta' => $this->generateCurrencyMeta($supplementalDataBundle),
             'Alpha3ToNumeric' => $this->generateAlpha3ToNumericMapping($numericCodesBundle, $this->currencyCodes),
-        );
+        ];
 
         $data['NumericToAlpha3'] = $this->generateNumericToAlpha3Mapping($data['Alpha3ToNumeric']);
 
@@ -181,14 +164,14 @@ class CurrencyDataGenerator extends AbstractDataGenerator
 
     private function generateNumericToAlpha3Mapping(array $alpha3ToNumericMapping)
     {
-        $numericToAlpha3Mapping = array();
+        $numericToAlpha3Mapping = [];
 
         foreach ($alpha3ToNumericMapping as $alpha3 => $numeric) {
             // Make sure that the mapping is stored as table and not as array
             $numeric = (string) $numeric;
 
             if (!isset($numericToAlpha3Mapping[$numeric])) {
-                $numericToAlpha3Mapping[$numeric] = array();
+                $numericToAlpha3Mapping[$numeric] = [];
             }
 
             $numericToAlpha3Mapping[$numeric][] = $alpha3;
