@@ -11,8 +11,8 @@
 
 namespace Symfony\Component\Intl\Data\Generator;
 
-use Symfony\Component\Intl\Data\Bundle\Compiler\GenrbCompiler;
-use Symfony\Component\Intl\Data\Bundle\Reader\BundleReaderInterface;
+use Symfony\Component\Intl\Data\Bundle\Compiler\BundleCompilerInterface;
+use Symfony\Component\Intl\Data\Bundle\Reader\BundleEntryReaderInterface;
 use Symfony\Component\Intl\Data\Util\ArrayAccessibleResourceBundle;
 use Symfony\Component\Intl\Data\Util\LocaleScanner;
 
@@ -27,40 +27,46 @@ use Symfony\Component\Intl\Data\Util\LocaleScanner;
  */
 class RegionDataGenerator extends AbstractDataGenerator
 {
-    const UNKNOWN_REGION_ID = 'ZZ';
-    const OUTLYING_OCEANIA_REGION_ID = 'QO';
-    const EUROPEAN_UNION_ID = 'EU';
-    const NETHERLANDS_ANTILLES_ID = 'AN';
-    const BOUVET_ISLAND_ID = 'BV';
-    const HEARD_MCDONALD_ISLANDS_ID = 'HM';
-    const CLIPPERTON_ISLAND_ID = 'CP';
-    const EUROZONE_ID = 'EZ';
-    const UNITED_NATIONS_ID = 'UN';
-
-    /**
-     * Regions excluded from generation.
-     */
-    private static $blacklist = array(
-        self::UNKNOWN_REGION_ID => true,
-        // Look like countries, but are sub-continents
-        self::OUTLYING_OCEANIA_REGION_ID => true,
-        self::EUROPEAN_UNION_ID => true,
-        self::EUROZONE_ID => true,
-        self::UNITED_NATIONS_ID => true,
-        // No longer exists
-        self::NETHERLANDS_ANTILLES_ID => true,
-        // Uninhabited islands
-        self::BOUVET_ISLAND_ID => true,
-        self::HEARD_MCDONALD_ISLANDS_ID => true,
-        self::CLIPPERTON_ISLAND_ID => true,
-    );
+    private static $blacklist = [
+        // Exceptional reservations
+        'AC' => true, // Ascension Island
+        'CP' => true, // Clipperton Island
+        'DG' => true, // Diego Garcia
+        'EA' => true, // Ceuta & Melilla
+        'EU' => true, // European Union
+        'EZ' => true, // Eurozone
+        'IC' => true, // Canary Islands
+        'TA' => true, // Tristan da Cunha
+        'UN' => true, // United Nations
+        // User-assigned
+        'QO' => true, // Outlying Oceania
+        'XA' => true, // Pseudo-Accents
+        'XB' => true, // Pseudo-Bidi
+        'XK' => true, // Kosovo
+        // Misc
+        'ZZ' => true, // Unknown Region
+    ];
 
     /**
      * Collects all available language codes.
      *
      * @var string[]
      */
-    private $regionCodes = array();
+    private $regionCodes = [];
+
+    public static function isValidCountryCode($region)
+    {
+        if (isset(self::$blacklist[$region])) {
+            return false;
+        }
+
+        // WORLD/CONTINENT/SUBCONTINENT/GROUPING
+        if (ctype_digit($region) || \is_int($region)) {
+            return false;
+        }
+
+        return true;
+    }
 
     /**
      * {@inheritdoc}
@@ -73,7 +79,7 @@ class RegionDataGenerator extends AbstractDataGenerator
     /**
      * {@inheritdoc}
      */
-    protected function compileTemporaryBundles(GenrbCompiler $compiler, $sourceDir, $tempDir)
+    protected function compileTemporaryBundles(BundleCompilerInterface $compiler, $sourceDir, $tempDir)
     {
         $compiler->compile($sourceDir.'/region', $tempDir);
     }
@@ -83,22 +89,22 @@ class RegionDataGenerator extends AbstractDataGenerator
      */
     protected function preGenerate()
     {
-        $this->regionCodes = array();
+        $this->regionCodes = [];
     }
 
     /**
      * {@inheritdoc}
      */
-    protected function generateDataForLocale(BundleReaderInterface $reader, $tempDir, $displayLocale)
+    protected function generateDataForLocale(BundleEntryReaderInterface $reader, $tempDir, $displayLocale)
     {
         $localeBundle = $reader->read($tempDir, $displayLocale);
 
         // isset() on \ResourceBundle returns true even if the value is null
         if (isset($localeBundle['Countries']) && null !== $localeBundle['Countries']) {
-            $data = array(
+            $data = [
                 'Version' => $localeBundle['Version'],
                 'Names' => $this->generateRegionNames($localeBundle),
-            );
+            ];
 
             $this->regionCodes = array_merge($this->regionCodes, array_keys($data['Names']));
 
@@ -109,14 +115,14 @@ class RegionDataGenerator extends AbstractDataGenerator
     /**
      * {@inheritdoc}
      */
-    protected function generateDataForRoot(BundleReaderInterface $reader, $tempDir)
+    protected function generateDataForRoot(BundleEntryReaderInterface $reader, $tempDir)
     {
     }
 
     /**
      * {@inheritdoc}
      */
-    protected function generateDataForMeta(BundleReaderInterface $reader, $tempDir)
+    protected function generateDataForMeta(BundleEntryReaderInterface $reader, $tempDir)
     {
         $rootBundle = $reader->read($tempDir, 'root');
 
@@ -124,10 +130,10 @@ class RegionDataGenerator extends AbstractDataGenerator
 
         sort($this->regionCodes);
 
-        return array(
+        return [
             'Version' => $rootBundle['Version'],
             'Regions' => $this->regionCodes,
-        );
+        ];
     }
 
     /**
@@ -136,15 +142,10 @@ class RegionDataGenerator extends AbstractDataGenerator
     protected function generateRegionNames(ArrayAccessibleResourceBundle $localeBundle)
     {
         $unfilteredRegionNames = iterator_to_array($localeBundle['Countries']);
-        $regionNames = array();
+        $regionNames = [];
 
         foreach ($unfilteredRegionNames as $region => $regionName) {
-            if (isset(self::$blacklist[$region])) {
-                continue;
-            }
-
-            // WORLD/CONTINENT/SUBCONTINENT/GROUPING
-            if (ctype_digit($region) || \is_int($region)) {
+            if (!self::isValidCountryCode($region)) {
                 continue;
             }
 
