@@ -11,8 +11,8 @@
 
 namespace Symfony\Component\Intl\Data\Generator;
 
-use Symfony\Component\Intl\Data\Bundle\Compiler\GenrbCompiler;
-use Symfony\Component\Intl\Data\Bundle\Reader\BundleReaderInterface;
+use Symfony\Component\Intl\Data\Bundle\Compiler\BundleCompilerInterface;
+use Symfony\Component\Intl\Data\Bundle\Reader\BundleEntryReaderInterface;
 use Symfony\Component\Intl\Data\Util\LocaleScanner;
 
 /**
@@ -24,12 +24,16 @@ use Symfony\Component\Intl\Data\Util\LocaleScanner;
  */
 class ScriptDataGenerator extends AbstractDataGenerator
 {
+    private static $blacklist = [
+        'Zzzz' => true, // Unknown Script
+    ];
+
     /**
      * Collects all available language codes.
      *
      * @var string[]
      */
-    private $scriptCodes = array();
+    private $scriptCodes = [];
 
     /**
      * {@inheritdoc}
@@ -42,7 +46,7 @@ class ScriptDataGenerator extends AbstractDataGenerator
     /**
      * {@inheritdoc}
      */
-    protected function compileTemporaryBundles(GenrbCompiler $compiler, $sourceDir, $tempDir)
+    protected function compileTemporaryBundles(BundleCompilerInterface $compiler, $sourceDir, $tempDir)
     {
         $compiler->compile($sourceDir.'/lang', $tempDir);
     }
@@ -52,22 +56,22 @@ class ScriptDataGenerator extends AbstractDataGenerator
      */
     protected function preGenerate()
     {
-        $this->scriptCodes = array();
+        $this->scriptCodes = [];
     }
 
     /**
      * {@inheritdoc}
      */
-    protected function generateDataForLocale(BundleReaderInterface $reader, $tempDir, $displayLocale)
+    protected function generateDataForLocale(BundleEntryReaderInterface $reader, $tempDir, $displayLocale)
     {
         $localeBundle = $reader->read($tempDir, $displayLocale);
 
         // isset() on \ResourceBundle returns true even if the value is null
         if (isset($localeBundle['Scripts']) && null !== $localeBundle['Scripts']) {
-            $data = array(
+            $data = [
                 'Version' => $localeBundle['Version'],
-                'Names' => iterator_to_array($localeBundle['Scripts']),
-            );
+                'Names' => array_diff_key(iterator_to_array($localeBundle['Scripts']), self::$blacklist),
+            ];
 
             $this->scriptCodes = array_merge($this->scriptCodes, array_keys($data['Names']));
 
@@ -78,14 +82,14 @@ class ScriptDataGenerator extends AbstractDataGenerator
     /**
      * {@inheritdoc}
      */
-    protected function generateDataForRoot(BundleReaderInterface $reader, $tempDir)
+    protected function generateDataForRoot(BundleEntryReaderInterface $reader, $tempDir)
     {
     }
 
     /**
      * {@inheritdoc}
      */
-    protected function generateDataForMeta(BundleReaderInterface $reader, $tempDir)
+    protected function generateDataForMeta(BundleEntryReaderInterface $reader, $tempDir)
     {
         $rootBundle = $reader->read($tempDir, 'root');
 
@@ -93,9 +97,9 @@ class ScriptDataGenerator extends AbstractDataGenerator
 
         sort($this->scriptCodes);
 
-        return array(
+        return [
             'Version' => $rootBundle['Version'],
             'Scripts' => $this->scriptCodes,
-        );
+        ];
     }
 }
