@@ -68,7 +68,7 @@ class LocaleScanner
     public function scanAliases($sourceDir)
     {
         $locales = $this->scanLocales($sourceDir);
-        $aliases = array();
+        $aliases = [];
 
         // Delete locales that are no aliases
         foreach ($locales as $locale) {
@@ -81,5 +81,25 @@ class LocaleScanner
         }
 
         return $aliases;
+    }
+
+    /**
+     * Returns all locale parents found in the given directory.
+     */
+    public function scanParents(string $sourceDir): array
+    {
+        $locales = $this->scanLocales($sourceDir);
+        $fallbacks = [];
+
+        foreach ($locales as $locale) {
+            $content = file_get_contents($sourceDir.'/'.$locale.'.txt');
+
+            // Aliases contain the text "%%PARENT" followed by the aliased locale
+            if (preg_match('/%%Parent{"([^"]+)"}/', $content, $matches)) {
+                $fallbacks[$locale] = $matches[1];
+            }
+        }
+
+        return $fallbacks;
     }
 }
