@@ -23,7 +23,7 @@ class QuarterTransformer extends Transformer
     /**
      * {@inheritdoc}
      */
-    public function format(\DateTime $dateTime, $length)
+    public function format(\DateTime $dateTime, int $length): string
     {
         $month = (int) $dateTime->format('n');
         $quarter = (int) floor(($month - 1) / 3) + 1;
@@ -34,7 +34,7 @@ class QuarterTransformer extends Transformer
             case 3:
                 return 'Q'.$quarter;
             default:
-                $map = array(1 => '1st quarter', 2 => '2nd quarter', 3 => '3rd quarter', 4 => '4th quarter');
+                $map = [1 => '1st quarter', 2 => '2nd quarter', 3 => '3rd quarter', 4 => '4th quarter'];
 
                 return $map[$quarter];
         }
@@ -43,7 +43,7 @@ class QuarterTransformer extends Transformer
     /**
      * {@inheritdoc}
      */
-    public function getReverseMatchingRegExp($length)
+    public function getReverseMatchingRegExp(int $length): string
     {
         switch ($length) {
             case 1:
@@ -59,8 +59,8 @@ class QuarterTransformer extends Transformer
     /**
      * {@inheritdoc}
      */
-    public function extractDateOptions($matched, $length)
+    public function extractDateOptions(string $matched, int $length): array
     {
-        return array();
+        return [];
     }
 }
