@@ -76,23 +76,23 @@ class IntlDateFormatter
     /**
      * Patterns used to format the date when no pattern is provided.
      */
-    private $defaultDateFormats = array(
+    private $defaultDateFormats = [
         self::NONE => '',
         self::FULL => 'EEEE, LLLL d, y',
         self::LONG => 'LLLL d, y',
         self::MEDIUM => 'LLL d, y',
         self::SHORT => 'M/d/yy',
-    );
+    ];
 
     /**
      * Patterns used to format the time when no pattern is provided.
      */
-    private $defaultTimeFormats = array(
+    private $defaultTimeFormats = [
         self::FULL => 'h:mm:ss a zzzz',
         self::LONG => 'h:mm:ss a z',
         self::MEDIUM => 'h:mm:ss a',
         self::SHORT => 'h:mm a',
-    );
+    ];
 
     private $datetype;
     private $timetype;
@@ -118,13 +118,13 @@ class IntlDateFormatter
     private $timeZoneId;
 
     /**
-     * @param string $locale   The locale code. The only currently supported locale is "en" (or null using the default locale, i.e. "en")
-     * @param int    $datetype Type of date formatting, one of the format type constants
-     * @param int    $timetype Type of time formatting, one of the format type constants
-     * @param mixed  $timezone Timezone identifier
-     * @param int    $calendar Calendar to use for formatting or parsing. The only currently
-     *                         supported value is IntlDateFormatter::GREGORIAN (or null using the default calendar, i.e. "GREGORIAN")
-     * @param string $pattern  Optional pattern to use when formatting
+     * @param string                                  $locale   The locale code. The only currently supported locale is "en" (or null using the default locale, i.e. "en")
+     * @param int|null                                $datetype Type of date formatting, one of the format type constants
+     * @param int|null                                $timetype Type of time formatting, one of the format type constants
+     * @param \IntlTimeZone|\DateTimeZone|string|null $timezone Timezone identifier
+     * @param int                                     $calendar Calendar to use for formatting or parsing. The only currently
+     *                                                          supported value is IntlDateFormatter::GREGORIAN (or null using the default calendar, i.e. "GREGORIAN")
+     * @param string|null                             $pattern  Optional pattern to use when formatting
      *
      * @see http://www.php.net/manual/en/intldateformatter.create.php
      * @see http://userguide.icu-project.org/formatparse/datetime
@@ -132,7 +132,7 @@ class IntlDateFormatter
      * @throws MethodArgumentValueNotImplementedException When $locale different than "en" or null is passed
      * @throws MethodArgumentValueNotImplementedException When $calendar different than GREGORIAN is passed
      */
-    public function __construct($locale, $datetype, $timetype, $timezone = null, $calendar = self::GREGORIAN, $pattern = null)
+    public function __construct(?string $locale, ?int $datetype, ?int $timetype, $timezone = null, ?int $calendar = self::GREGORIAN, string $pattern = null)
     {
         if ('en' !== $locale && null !== $locale) {
             throw new MethodArgumentValueNotImplementedException(__METHOD__, 'locale', $locale, 'Only the locale "en" is supported');
@@ -142,8 +142,8 @@ class IntlDateFormatter
             throw new MethodArgumentValueNotImplementedException(__METHOD__, 'calendar', $calendar, 'Only the GREGORIAN calendar is supported');
         }
 
-        $this->datetype = $datetype;
-        $this->timetype = $timetype;
+        $this->datetype = null !== $datetype ? $datetype : self::FULL;
+        $this->timetype = null !== $timetype ? $timetype : self::FULL;
 
         $this->setPattern($pattern);
         $this->setTimeZone($timezone);
@@ -152,13 +152,13 @@ class IntlDateFormatter
     /**
      * Static constructor.
      *
-     * @param string $locale   The locale code. The only currently supported locale is "en" (or null using the default locale, i.e. "en")
-     * @param int    $datetype Type of date formatting, one of the format type constants
-     * @param int    $timetype Type of time formatting, one of the format type constants
-     * @param string $timezone Timezone identifier
-     * @param int    $calendar Calendar to use for formatting or parsing; default is Gregorian
-     *                         One of the calendar constants
-     * @param string $pattern  Optional pattern to use when formatting
+     * @param string                                  $locale   The locale code. The only currently supported locale is "en" (or null using the default locale, i.e. "en")
+     * @param int|null                                $datetype Type of date formatting, one of the format type constants
+     * @param int|null                                $timetype Type of time formatting, one of the format type constants
+     * @param \IntlTimeZone|\DateTimeZone|string|null $timezone Timezone identifier
+     * @param int                                     $calendar Calendar to use for formatting or parsing; default is Gregorian
+     *                                                          One of the calendar constants
+     * @param string|null                             $pattern  Optional pattern to use when formatting
      *
      * @return self
      *
@@ -176,8 +176,7 @@ class IntlDateFormatter
     /**
      * Format the date/time value (timestamp) as a string.
      *
-     * @param int|\DateTime $timestamp The timestamp to format. \DateTime objects
-     *                                 are supported as of PHP 5.3.4.
+     * @param int|\DateTime $timestamp The timestamp to format
      *
      * @return string|bool The formatted value or false if formatting failed
      *
@@ -197,10 +196,7 @@ class IntlDateFormatter
         // behave like the intl extension
         $argumentError = null;
         if (!\is_int($timestamp) && !$timestamp instanceof \DateTime) {
-            $argumentError = 'datefmt_format: takes either an array or an integer timestamp value or a DateTime object';
-            if (\PHP_VERSION_ID >= 50500 || (\extension_loaded('intl') && method_exists('IntlDateFormatter', 'setTimeZone'))) {
-                $argumentError = sprintf('datefmt_format: string \'%s\' is not numeric, which would be required for it to be a valid date', $timestamp);
-            }
+            $argumentError = sprintf('datefmt_format: string \'%s\' is not numeric, which would be required for it to be a valid date', $timestamp);
         }
 
         if (null !== $argumentError) {
@@ -359,10 +355,7 @@ class IntlDateFormatter
             return $this->timeZoneId;
         }
 
-        // In PHP 5.5 default timezone depends on `date_default_timezone_get()` method
-        if (\PHP_VERSION_ID >= 50500 || (\extension_loaded('intl') && method_exists('IntlDateFormatter', 'setTimeZone'))) {
-            return date_default_timezone_get();
-        }
+        return date_default_timezone_get();
     }
 
     /**
@@ -492,7 +485,7 @@ class IntlDateFormatter
     /**
      * Set the formatter's pattern.
      *
-     * @param string $pattern A pattern string in conformance with the ICU IntlDateFormatter documentation
+     * @param string|null $pattern A pattern string in conformance with the ICU IntlDateFormatter documentation
      *
      * @return bool true on success or false on failure
      *
@@ -513,9 +506,9 @@ class IntlDateFormatter
     /**
      * Set the formatter's timezone identifier.
      *
-     * @param string $timeZoneId The time zone ID string of the time zone to use.
-     *                           If NULL or the empty string, the default time zone for the
-     *                           runtime is used.
+     * @param string|null $timeZoneId The time zone ID string of the time zone to use.
+     *                                If NULL or the empty string, the default time zone for the
+     *                                runtime is used.
      *
      * @return bool true on success or false on failure
      *
@@ -524,16 +517,7 @@ class IntlDateFormatter
     public function setTimeZoneId($timeZoneId)
     {
         if (null === $timeZoneId) {
-            // In PHP 5.5 if $timeZoneId is null it fallbacks to `date_default_timezone_get()` method
-            if (\PHP_VERSION_ID >= 50500 || (\extension_loaded('intl') && method_exists('IntlDateFormatter', 'setTimeZone'))) {
-                $timeZoneId = date_default_timezone_get();
-            } else {
-                // TODO: changes were made to ext/intl in PHP 5.4.4 release that need to be investigated since it will
-                // use ini's date.timezone when the time zone is not provided. As a not well tested workaround, uses UTC.
-                // See the first two items of the commit message for more information:
-                // https://github.com/php/php-src/commit/eb346ef0f419b90739aadfb6cc7b7436c5b521d9
-                $timeZoneId = getenv('TZ') ?: 'UTC';
-            }
+            $timeZoneId = date_default_timezone_get();
 
             $this->uninitializedTimeZoneId = true;
         }
@@ -544,7 +528,7 @@ class IntlDateFormatter
         // Get an Etc/GMT time zone that is accepted for \DateTimeZone
         if ('GMT' !== $timeZoneId && 0 === strpos($timeZoneId, 'GMT')) {
             try {
-                $timeZoneId = DateFormat\TimeZoneTransformer::getEtcTimeZoneId($timeZoneId);
+                $timeZoneId = DateFormat\TimezoneTransformer::getEtcTimeZoneId($timeZoneId);
             } catch (\InvalidArgumentException $e) {
                 // Does nothing, will fallback to UTC
             }
@@ -556,11 +540,7 @@ class IntlDateFormatter
                 $timeZone = $this->getTimeZoneId();
             }
         } catch (\Exception $e) {
-            if (\PHP_VERSION_ID >= 50500 || (\extension_loaded('intl') && method_exists('IntlDateFormatter', 'setTimeZone'))) {
-                $timeZoneId = $timeZone = $this->getTimeZoneId();
-            } else {
-                $timeZoneId = 'UTC';
-            }
+            $timeZoneId = $timeZone = $this->getTimeZoneId();
             $this->dateTimeZone = new \DateTimeZone($timeZoneId);
         }
 
@@ -572,7 +552,7 @@ class IntlDateFormatter
     /**
      * This method was added in PHP 5.5 as replacement for `setTimeZoneId()`.
      *
-     * @param mixed $timeZone
+     * @param \IntlTimeZone|\DateTimeZone|string|null $timeZone
      *
      * @return bool true on success or false on failure
      *
@@ -620,7 +600,7 @@ class IntlDateFormatter
      */
     protected function getDefaultPattern()
     {
-        $patternParts = array();
+        $patternParts = [];
         if (self::NONE !== $this->datetype) {
             $patternParts[] = $this->defaultDateFormats[$this->datetype];
         }
