@@ -11,12 +11,12 @@
 
 namespace Symfony\Component\Intl\NumberFormatter;
 
+use Symfony\Component\Intl\Currencies;
 use Symfony\Component\Intl\Exception\MethodArgumentNotImplementedException;
 use Symfony\Component\Intl\Exception\MethodArgumentValueNotImplementedException;
 use Symfony\Component\Intl\Exception\MethodNotImplementedException;
 use Symfony\Component\Intl\Exception\NotImplementedException;
 use Symfony\Component\Intl\Globals\IntlGlobals;
-use Symfony\Component\Intl\Intl;
 use Symfony\Component\Intl\Locale\Locale;
 
 /**
@@ -151,40 +151,40 @@ class NumberFormatter
     /**
      * Default values for the en locale.
      */
-    private $attributes = array(
+    private $attributes = [
         self::FRACTION_DIGITS => 0,
         self::GROUPING_USED => 1,
         self::ROUNDING_MODE => self::ROUND_HALFEVEN,
-    );
+    ];
 
     /**
      * Holds the initialized attributes code.
      */
-    private $initializedAttributes = array();
+    private $initializedAttributes = [];
 
     /**
      * The supported styles to the constructor $styles argument.
      */
-    private static $supportedStyles = array(
+    private static $supportedStyles = [
         'CURRENCY' => self::CURRENCY,
         'DECIMAL' => self::DECIMAL,
-    );
+    ];
 
     /**
      * Supported attributes to the setAttribute() $attr argument.
      */
-    private static $supportedAttributes = array(
+    private static $supportedAttributes = [
         'FRACTION_DIGITS' => self::FRACTION_DIGITS,
         'GROUPING_USED' => self::GROUPING_USED,
         'ROUNDING_MODE' => self::ROUNDING_MODE,
-    );
+    ];
 
     /**
      * The available rounding modes for setAttribute() usage with
      * NumberFormatter::ROUNDING_MODE. NumberFormatter::ROUND_DOWN
      * and NumberFormatter::ROUND_UP does not have a PHP only equivalent.
      */
-    private static $roundingModes = array(
+    private static $roundingModes = [
         'ROUND_HALFEVEN' => self::ROUND_HALFEVEN,
         'ROUND_HALFDOWN' => self::ROUND_HALFDOWN,
         'ROUND_HALFUP' => self::ROUND_HALFUP,
@@ -192,7 +192,7 @@ class NumberFormatter
         'ROUND_FLOOR' => self::ROUND_FLOOR,
         'ROUND_DOWN' => self::ROUND_DOWN,
         'ROUND_UP' => self::ROUND_UP,
-    );
+    ];
 
     /**
      * The mapping between NumberFormatter rounding modes to the available
@@ -200,23 +200,23 @@ class NumberFormatter
      *
      * @see http://www.php.net/manual/en/function.round.php
      */
-    private static $phpRoundingMap = array(
+    private static $phpRoundingMap = [
         self::ROUND_HALFDOWN => \PHP_ROUND_HALF_DOWN,
         self::ROUND_HALFEVEN => \PHP_ROUND_HALF_EVEN,
         self::ROUND_HALFUP => \PHP_ROUND_HALF_UP,
-    );
+    ];
 
     /**
      * The list of supported rounding modes which aren't available modes in
      * PHP's round() function, but there's an equivalent. Keys are rounding
      * modes, values does not matter.
      */
-    private static $customRoundingList = array(
+    private static $customRoundingList = [
         self::ROUND_CEILING => true,
         self::ROUND_FLOOR => true,
         self::ROUND_DOWN => true,
         self::ROUND_UP => true,
-    );
+    ];
 
     /**
      * The maximum value of the integer type in 32 bit platforms.
@@ -230,15 +230,15 @@ class NumberFormatter
      */
     private static $int64Max = 9223372036854775807;
 
-    private static $enSymbols = array(
-        self::DECIMAL => array('.', ',', ';', '%', '0', '#', '-', '+', '¤', '¤¤', '.', 'E', '‰', '*', '∞', 'NaN', '@', ','),
-        self::CURRENCY => array('.', ',', ';', '%', '0', '#', '-', '+', '¤', '¤¤', '.', 'E', '‰', '*', '∞', 'NaN', '@', ','),
-    );
+    private static $enSymbols = [
+        self::DECIMAL => ['.', ',', ';', '%', '0', '#', '-', '+', '¤', '¤¤', '.', 'E', '‰', '*', '∞', 'NaN', '@', ','],
+        self::CURRENCY => ['.', ',', ';', '%', '0', '#', '-', '+', '¤', '¤¤', '.', 'E', '‰', '*', '∞', 'NaN', '@', ','],
+    ];
 
-    private static $enTextAttributes = array(
-        self::DECIMAL => array('', '', '-', '', ' ', '', ''),
-        self::CURRENCY => array('¤', '', '-¤', '', ' ', ''),
-    );
+    private static $enTextAttributes = [
+        self::DECIMAL => ['', '', '-', '', ' ', 'XXX', ''],
+        self::CURRENCY => ['¤', '', '-¤', '', ' ', 'XXX'],
+    ];
 
     /**
      * @param string $locale  The locale code. The only currently supported locale is "en" (or null using the default locale, i.e. "en")
@@ -257,7 +257,7 @@ class NumberFormatter
      * @throws MethodArgumentValueNotImplementedException When the $style is not supported
      * @throws MethodArgumentNotImplementedException      When the pattern value is different than null
      */
-    public function __construct($locale = 'en', $style = null, $pattern = null)
+    public function __construct(?string $locale = 'en', int $style = null, $pattern = null)
     {
         if ('en' !== $locale && null !== $locale) {
             throw new MethodArgumentValueNotImplementedException(__METHOD__, 'locale', $locale, 'Only the locale "en" is supported');
@@ -318,8 +318,8 @@ class NumberFormatter
             return $this->format($value);
         }
 
-        $symbol = Intl::getCurrencyBundle()->getCurrencySymbol($currency, 'en');
-        $fractionDigits = Intl::getCurrencyBundle()->getFractionDigits($currency);
+        $symbol = Currencies::getSymbol($currency, 'en');
+        $fractionDigits = Currencies::getFractionDigits($currency);
 
         $value = $this->roundCurrency($value, $currency);
 
@@ -331,7 +331,8 @@ class NumberFormatter
 
         $value = $this->formatNumber($value, $fractionDigits);
 
-        $ret = $symbol.$value;
+        // There's a non-breaking space after the currency code (i.e. CRC 100), but not if the currency has a symbol (i.e. £100).
+        $ret = $symbol.(mb_strlen($symbol, 'UTF-8') > 2 ? "\xc2\xa0" : '').$value;
 
         return $negative ? '-'.$ret : $ret;
     }
@@ -459,7 +460,7 @@ class NumberFormatter
      */
     public function getSymbol($attr)
     {
-        return array_key_exists($this->style, self::$enSymbols) && array_key_exists($attr, self::$enSymbols[$this->style]) ? self::$enSymbols[$this->style][$attr] : false;
+        return \array_key_exists($this->style, self::$enSymbols) && \array_key_exists($attr, self::$enSymbols[$this->style]) ? self::$enSymbols[$this->style][$attr] : false;
     }
 
     /**
@@ -473,7 +474,7 @@ class NumberFormatter
      */
     public function getTextAttribute($attr)
     {
-        return array_key_exists($this->style, self::$enTextAttributes) && array_key_exists($attr, self::$enTextAttributes[$this->style]) ? self::$enTextAttributes[$this->style][$attr] : false;
+        return \array_key_exists($this->style, self::$enTextAttributes) && \array_key_exists($attr, self::$enTextAttributes[$this->style]) ? self::$enTextAttributes[$this->style][$attr] : false;
     }
 
     /**
@@ -483,7 +484,7 @@ class NumberFormatter
      * @param string $currency Parameter to receive the currency name (reference)
      * @param int    $position Offset to begin the parsing on return this value will hold the offset at which the parsing ended
      *
-     * @return bool|string The parsed numeric value of false on error
+     * @return bool|string The parsed numeric value or false on error
      *
      * @see http://www.php.net/manual/en/numberformatter.parsecurrency.php
      *
@@ -501,7 +502,7 @@ class NumberFormatter
      * @param int    $type     Type of the formatting, one of the format type constants. NumberFormatter::TYPE_DOUBLE by default.
      * @param int    $position Offset to begin the parsing on return this value will hold the offset at which the parsing ended
      *
-     * @return int|float|false The parsed value of false on error
+     * @return int|float|false The parsed value or false on error
      *
      * @see http://www.php.net/manual/en/numberformatter.parse.php
      */
@@ -513,17 +514,20 @@ class NumberFormatter
             return false;
         }
 
-        $groupSep = $this->getAttribute(self::GROUPING_USED) ? ',' : '';
-
-        // Any string before the numeric value causes error in the parsing
-        if (preg_match("/^-?(?:\.\d++|([\d{$groupSep}]++)(?:\.\d*+)?)/", $value, $matches)) {
+        // Any invalid number at the end of the string is removed.
+        // Only numbers and the fraction separator is expected in the string.
+        // If grouping is used, grouping separator also becomes a valid character.
+        $groupingMatch = $this->getAttribute(self::GROUPING_USED) ? '|(?P<grouping>\d++(,{1}\d+)++(\.\d*+)?)' : '';
+        if (preg_match("/^-?(?:\.\d++{$groupingMatch}|\d++(\.\d*+)?)/", $value, $matches)) {
             $value = $matches[0];
             $position = \strlen($value);
-            if ($error = $groupSep && isset($matches[1]) && !preg_match('/^\d{1,3}+(?:(?:,\d{3})++|\d*+)$/', $matches[1])) {
-                $position -= \strlen(preg_replace('/^\d{1,3}+(?:(?:,\d++)++|\d*+)/', '', $matches[1]));
+            // value is not valid if grouping is used, but digits are not grouped in groups of three
+            if ($error = isset($matches['grouping']) && !preg_match('/^-?(?:\d{1,3}+)?(?:(?:,\d{3})++|\d*+)(?:\.\d*+)?$/', $value)) {
+                // the position on error is 0 for positive and 1 for negative numbers
+                $position = 0 === strpos($value, '-') ? 1 : 0;
             }
         } else {
-            $error = 1;
+            $error = true;
             $position = 0;
         }
 
@@ -585,6 +589,10 @@ class NumberFormatter
 
         if (self::$supportedAttributes['FRACTION_DIGITS'] == $attr) {
             $value = $this->normalizeFractionDigitsValue($value);
+            if ($value < 0) {
+                // ignore negative values but do not raise an error
+                return true;
+            }
         }
 
         $this->attributes[$attr] = $value;
@@ -630,8 +638,8 @@ class NumberFormatter
     /**
      * Not supported. Set a text attribute.
      *
-     * @param int $attr  An attribute specifier, one of the text attribute constants
-     * @param int $value The attribute value
+     * @param int    $attr  An attribute specifier, one of the text attribute constants
+     * @param string $value The attribute value
      *
      * @return bool true on success or false on failure
      *
@@ -674,8 +682,8 @@ class NumberFormatter
      */
     private function roundCurrency($value, $currency)
     {
-        $fractionDigits = Intl::getCurrencyBundle()->getFractionDigits($currency);
-        $roundingIncrement = Intl::getCurrencyBundle()->getRoundingIncrement($currency);
+        $fractionDigits = Currencies::getFractionDigits($currency);
+        $roundingIncrement = Currencies::getRoundingIncrement($currency);
 
         // Round with the formatter rounding mode
         $value = $this->round($value, $fractionDigits);
@@ -831,29 +839,8 @@ class NumberFormatter
             return false;
         }
 
-        if (PHP_INT_SIZE !== 8 && ($value > self::$int32Max || $value <= -self::$int32Max - 1)) {
-            // Bug #59597 was fixed on PHP 5.3.14 and 5.4.4
-            // The negative PHP_INT_MAX was being converted to float
-            if (
-                $value == -self::$int32Max - 1 &&
-                ((\PHP_VERSION_ID < 50400 && \PHP_VERSION_ID >= 50314) || \PHP_VERSION_ID >= 50404 || (\extension_loaded('intl') && method_exists('IntlDateFormatter', 'setTimeZone')))
-            ) {
-                return (int) $value;
-            }
-
+        if (PHP_INT_SIZE !== 8 && ($value > self::$int32Max || $value < -self::$int32Max - 1)) {
             return (float) $value;
-        }
-
-        if (PHP_INT_SIZE === 8) {
-            // Bug #59597 was fixed on PHP 5.3.14 and 5.4.4
-            // A 32 bit integer was being generated instead of a 64 bit integer
-            if (
-                  ($value > self::$int32Max || $value < -self::$int32Max - 1) &&
-                  (\PHP_VERSION_ID < 50314 || (\PHP_VERSION_ID >= 50400 && \PHP_VERSION_ID < 50404)) &&
-                  !(\extension_loaded('intl') && method_exists('IntlDateFormatter', 'setTimeZone'))
-            ) {
-                $value = (-2147483648 - ($value % -2147483648)) * ($value / abs($value));
-            }
         }
 
         return (int) $value;
@@ -889,8 +876,7 @@ class NumberFormatter
     }
 
     /**
-     * Returns the normalized value for the FRACTION_DIGITS attribute. The value is converted to int and if negative,
-     * the returned value will be 0.
+     * Returns the normalized value for the FRACTION_DIGITS attribute.
      *
      * @param mixed $value The value to be normalized
      *
@@ -898,8 +884,6 @@ class NumberFormatter
      */
     private function normalizeFractionDigitsValue($value)
     {
-        $value = (int) $value;
-
-        return (0 > $value) ? 0 : $value;
+        return (int) $value;
     }
 }
