@@ -50,51 +50,51 @@ class TextBundleWriterTest extends TestCase
 
     public function testWrite()
     {
-        $this->writer->write($this->directory, 'en', array(
-            'Entry1' => array(
-                'Array' => array('foo', 'bar', array('Key' => 'value')),
+        $this->writer->write($this->directory, 'en', [
+            'Entry1' => [
+                'Array' => ['foo', 'bar', ['Key' => 'value']],
                 'Integer' => 5,
-                'IntVector' => array(0, 1, 2, 3),
-                'NotAnIntVector' => array(0 => 0, 2 => 1, 1 => 2, 3 => 3),
-                'IntVectorWithStringKeys' => array('a' => 0, 'b' => 1, 'c' => 2),
-                'TableWithIntKeys' => array(0 => 0, 1 => 1, 3 => 3),
+                'IntVector' => [0, 1, 2, 3],
+                'NotAnIntVector' => [0 => 0, 2 => 1, 1 => 2, 3 => 3],
+                'IntVectorWithStringKeys' => ['a' => 0, 'b' => 1, 'c' => 2],
+                'TableWithIntKeys' => [0 => 0, 1 => 1, 3 => 3],
                 'FalseBoolean' => false,
                 'TrueBoolean' => true,
                 'Null' => null,
                 'Float' => 1.23,
-            ),
+            ],
             'Entry2' => 'String',
-        ));
+        ]);
 
         $this->assertFileEquals(__DIR__.'/Fixtures/en.txt', $this->directory.'/en.txt');
     }
 
     public function testWriteTraversable()
     {
-        $this->writer->write($this->directory, 'en', new \ArrayIterator(array(
-            'Entry1' => new \ArrayIterator(array(
-                'Array' => array('foo', 'bar', array('Key' => 'value')),
+        $this->writer->write($this->directory, 'en', new \ArrayIterator([
+            'Entry1' => new \ArrayIterator([
+                'Array' => ['foo', 'bar', ['Key' => 'value']],
                 'Integer' => 5,
-                'IntVector' => array(0, 1, 2, 3),
-                'NotAnIntVector' => array(0 => 0, 2 => 1, 1 => 2, 3 => 3),
-                'IntVectorWithStringKeys' => array('a' => 0, 'b' => 1, 'c' => 2),
-                'TableWithIntKeys' => array(0 => 0, 1 => 1, 3 => 3),
+                'IntVector' => [0, 1, 2, 3],
+                'NotAnIntVector' => [0 => 0, 2 => 1, 1 => 2, 3 => 3],
+                'IntVectorWithStringKeys' => ['a' => 0, 'b' => 1, 'c' => 2],
+                'TableWithIntKeys' => [0 => 0, 1 => 1, 3 => 3],
                 'FalseBoolean' => false,
                 'TrueBoolean' => true,
                 'Null' => null,
                 'Float' => 1.23,
-            )),
+            ]),
             'Entry2' => 'String',
-        )));
+        ]));
 
         $this->assertFileEquals(__DIR__.'/Fixtures/en.txt', $this->directory.'/en.txt');
     }
 
     public function testWriteNoFallback()
     {
-        $data = array(
+        $data = [
             'Entry' => 'Value',
-        );
+        ];
 
         $this->writer->write($this->directory, 'en_nofallback', $data, $fallback = false);
 
@@ -103,13 +103,13 @@ class TextBundleWriterTest extends TestCase
 
     public function testEscapeKeysIfNecessary()
     {
-        $this->writer->write($this->directory, 'escaped', array(
+        $this->writer->write($this->directory, 'escaped', [
             // Keys with colons must be escaped, otherwise the part after the
             // colon is interpreted as resource type
             'EntryWith:Colon' => 'Value',
             // Keys with spaces must be escaped
             'Entry With Spaces' => 'Value',
-        ));
+        ]);
 
         $this->assertFileEquals(__DIR__.'/Fixtures/escaped.txt', $this->directory.'/escaped.txt');
     }
