@@ -42,10 +42,13 @@ class LocaleScannerTest extends TestCase
 
         $this->filesystem->touch($this->directory.'/en.txt');
         $this->filesystem->touch($this->directory.'/en_alias.txt');
+        $this->filesystem->touch($this->directory.'/en_child.txt');
         $this->filesystem->touch($this->directory.'/de.txt');
         $this->filesystem->touch($this->directory.'/de_alias.txt');
+        $this->filesystem->touch($this->directory.'/de_child.txt');
         $this->filesystem->touch($this->directory.'/fr.txt');
         $this->filesystem->touch($this->directory.'/fr_alias.txt');
+        $this->filesystem->touch($this->directory.'/fr_child.txt');
         $this->filesystem->touch($this->directory.'/root.txt');
         $this->filesystem->touch($this->directory.'/supplementalData.txt');
         $this->filesystem->touch($this->directory.'/supplementaldata.txt');
@@ -54,6 +57,9 @@ class LocaleScannerTest extends TestCase
         file_put_contents($this->directory.'/en_alias.txt', 'en_alias{"%%ALIAS"{"en"}}');
         file_put_contents($this->directory.'/de_alias.txt', 'de_alias{"%%ALIAS"{"de"}}');
         file_put_contents($this->directory.'/fr_alias.txt', 'fr_alias{"%%ALIAS"{"fr"}}');
+        file_put_contents($this->directory.'/en_child.txt', 'en_GB{%%Parent{"en"}}');
+        file_put_contents($this->directory.'/de_child.txt', 'en_GB{%%Parent{"de"}}');
+        file_put_contents($this->directory.'/fr_child.txt', 'en_GB{%%Parent{"fr"}}');
     }
 
     protected function tearDown()
@@ -63,15 +69,22 @@ class LocaleScannerTest extends TestCase
 
     public function testScanLocales()
     {
-        $sortedLocales = array('de', 'de_alias', 'en', 'en_alias', 'fr', 'fr_alias');
+        $sortedLocales = ['de', 'de_alias', 'de_child', 'en', 'en_alias', 'en_child', 'fr', 'fr_alias', 'fr_child'];
 
         $this->assertSame($sortedLocales, $this->scanner->scanLocales($this->directory));
     }
 
     public function testScanAliases()
     {
-        $sortedAliases = array('de_alias' => 'de', 'en_alias' => 'en', 'fr_alias' => 'fr');
+        $sortedAliases = ['de_alias' => 'de', 'en_alias' => 'en', 'fr_alias' => 'fr'];
 
         $this->assertSame($sortedAliases, $this->scanner->scanAliases($this->directory));
+    }
+
+    public function testScanParents()
+    {
+        $sortedParents = ['de_child' => 'de', 'en_child' => 'en', 'fr_child' => 'fr'];
+
+        $this->assertSame($sortedParents, $this->scanner->scanParents($this->directory));
     }
 }
