@@ -11,6 +11,8 @@
 
 namespace Symfony\Component\Intl\Tests\Locale;
 
+use Symfony\Component\Intl\Locale\Locale;
+
 class LocaleTest extends AbstractLocaleTest
 {
     /**
@@ -21,16 +23,27 @@ class LocaleTest extends AbstractLocaleTest
         $this->call('acceptFromHttp', 'pt-br,en-us;q=0.7,en;q=0.5');
     }
 
+    public function testCanonicalize()
+    {
+        $this->assertSame('en', $this->call('canonicalize', ''));
+        $this->assertSame('en', $this->call('canonicalize', '.utf8'));
+        $this->assertSame('fr_FR', $this->call('canonicalize', 'FR-fr'));
+        $this->assertSame('fr_FR', $this->call('canonicalize', 'FR-fr.utf8'));
+        $this->assertSame('uz_Latn', $this->call('canonicalize', 'UZ-lATN'));
+        $this->assertSame('uz_Cyrl_UZ', $this->call('canonicalize', 'UZ-cYRL-uz'));
+        $this->assertSame('123', $this->call('canonicalize', 123));
+    }
+
     /**
      * @expectedException \Symfony\Component\Intl\Exception\MethodNotImplementedException
      */
     public function testComposeLocale()
     {
-        $subtags = array(
+        $subtags = [
             'language' => 'pt',
             'script' => 'Latn',
             'region' => 'BR',
-        );
+        ];
         $this->call('composeLocale', $subtags);
     }
 
@@ -127,10 +140,10 @@ class LocaleTest extends AbstractLocaleTest
      */
     public function testLookup()
     {
-        $langtag = array(
+        $langtag = [
             'pt-Latn-BR',
             'pt-BR',
-        );
+        ];
         $this->call('lookup', $langtag, 'pt-BR-x-priv1');
     }
 
@@ -161,6 +174,6 @@ class LocaleTest extends AbstractLocaleTest
     {
         $args = \array_slice(\func_get_args(), 1);
 
-        return \call_user_func_array(array('Symfony\Component\Intl\Locale\Locale', $methodName), $args);
+        return Locale::{$methodName}(...$args);
     }
 }
