@@ -18,22 +18,22 @@ class LocaleTest extends TestCase
 {
     public function provideGetFallbackTests()
     {
-        $tests = array(
-            array('sl_Latn_IT', 'sl_Latn_IT_nedis'),
-            array('sl_Latn', 'sl_Latn_IT'),
-            array('fr', 'fr_FR'),
-            array('fr', 'fr-FR'),
-            array('en', 'fr'),
-            array('root', 'en'),
-            array(null, 'root'),
-        );
+        $tests = [
+            ['sl_Latn_IT', 'sl_Latn_IT_nedis'],
+            ['sl_Latn', 'sl_Latn_IT'],
+            ['fr', 'fr_FR'],
+            ['fr', 'fr-FR'],
+            ['en', 'fr'],
+            ['root', 'en'],
+            [null, 'root'],
+        ];
 
         if (\function_exists('locale_parse')) {
-            $tests[] = array('sl_Latn_IT', 'sl-Latn-IT-nedis');
-            $tests[] = array('sl_Latn', 'sl-Latn_IT');
+            $tests[] = ['sl_Latn_IT', 'sl-Latn-IT-nedis'];
+            $tests[] = ['sl_Latn', 'sl-Latn_IT'];
         } else {
-            $tests[] = array('sl-Latn-IT', 'sl-Latn-IT-nedis');
-            $tests[] = array('sl-Latn', 'sl-Latn-IT');
+            $tests[] = ['sl-Latn-IT', 'sl-Latn-IT-nedis'];
+            $tests[] = ['sl-Latn', 'sl-Latn-IT'];
         }
 
         return $tests;
@@ -45,5 +45,29 @@ class LocaleTest extends TestCase
     public function testGetFallback($expected, $locale)
     {
         $this->assertSame($expected, Locale::getFallback($locale));
+    }
+
+    public function testNoDefaultFallback()
+    {
+        $prev = Locale::getDefaultFallback();
+        Locale::setDefaultFallback(null);
+
+        $this->assertSame('nl', Locale::getFallback('nl_NL'));
+        $this->assertNull(Locale::getFallback('nl'));
+        $this->assertNull(Locale::getFallback('root'));
+
+        Locale::setDefaultFallback($prev);
+    }
+
+    public function testDefaultRootFallback()
+    {
+        $prev = Locale::getDefaultFallback();
+        Locale::setDefaultFallback('root');
+
+        $this->assertSame('nl', Locale::getFallback('nl_NL'));
+        $this->assertSame('root', Locale::getFallback('nl'));
+        $this->assertNull(Locale::getFallback('root'));
+
+        Locale::setDefaultFallback($prev);
     }
 }
