@@ -29,7 +29,7 @@ class BufferedBundleReader implements BundleReaderInterface
      * @param BundleReaderInterface $reader     The reader to buffer
      * @param int                   $bufferSize The number of entries to store in the buffer
      */
-    public function __construct(BundleReaderInterface $reader, $bufferSize)
+    public function __construct(BundleReaderInterface $reader, int $bufferSize)
     {
         $this->reader = $reader;
         $this->buffer = new RingBuffer($bufferSize);
