@@ -25,7 +25,7 @@ class RecursiveArrayAccess
         foreach ($indices as $index) {
             // Use array_key_exists() for arrays, isset() otherwise
             if (\is_array($array)) {
-                if (array_key_exists($index, $array)) {
+                if (\array_key_exists($index, $array)) {
                     $array = $array[$index];
                     continue;
                 }
