@@ -23,9 +23,9 @@ class DayOfYearTransformer extends Transformer
     /**
      * {@inheritdoc}
      */
-    public function format(\DateTime $dateTime, $length)
+    public function format(\DateTime $dateTime, int $length): string
     {
-        $dayOfYear = $dateTime->format('z') + 1;
+        $dayOfYear = (int) $dateTime->format('z') + 1;
 
         return $this->padLeft($dayOfYear, $length);
     }
@@ -33,7 +33,7 @@ class DayOfYearTransformer extends Transformer
     /**
      * {@inheritdoc}
      */
-    public function getReverseMatchingRegExp($length)
+    public function getReverseMatchingRegExp(int $length): string
     {
         return '\d{'.$length.'}';
     }
@@ -41,8 +41,8 @@ class DayOfYearTransformer extends Transformer
     /**
      * {@inheritdoc}
      */
-    public function extractDateOptions($matched, $length)
+    public function extractDateOptions(string $matched, int $length): array
     {
-        return array();
+        return [];
     }
 }
