@@ -23,7 +23,7 @@ class DayTransformer extends Transformer
     /**
      * {@inheritdoc}
      */
-    public function format(\DateTime $dateTime, $length)
+    public function format(\DateTime $dateTime, int $length): string
     {
         return $this->padLeft($dateTime->format('j'), $length);
     }
@@ -31,7 +31,7 @@ class DayTransformer extends Transformer
     /**
      * {@inheritdoc}
      */
-    public function getReverseMatchingRegExp($length)
+    public function getReverseMatchingRegExp(int $length): string
     {
         return 1 === $length ? '\d{1,2}' : '\d{'.$length.'}';
     }
@@ -39,10 +39,10 @@ class DayTransformer extends Transformer
     /**
      * {@inheritdoc}
      */
-    public function extractDateOptions($matched, $length)
+    public function extractDateOptions(string $matched, int $length): array
     {
-        return array(
+        return [
             'day' => (int) $matched,
-        );
+        ];
     }
 }
