@@ -23,7 +23,7 @@ class Hour1200Transformer extends HourTransformer
     /**
      * {@inheritdoc}
      */
-    public function format(\DateTime $dateTime, $length)
+    public function format(\DateTime $dateTime, int $length): string
     {
         $hourOfDay = $dateTime->format('g');
         $hourOfDay = '12' == $hourOfDay ? '0' : $hourOfDay;
@@ -34,7 +34,7 @@ class Hour1200Transformer extends HourTransformer
     /**
      * {@inheritdoc}
      */
-    public function normalizeHour($hour, $marker = null)
+    public function normalizeHour(int $hour, string $marker = null): int
     {
         if ('PM' === $marker) {
             $hour += 12;
@@ -46,7 +46,7 @@ class Hour1200Transformer extends HourTransformer
     /**
      * {@inheritdoc}
      */
-    public function getReverseMatchingRegExp($length)
+    public function getReverseMatchingRegExp(int $length): string
     {
         return '\d{1,2}';
     }
@@ -54,11 +54,11 @@ class Hour1200Transformer extends HourTransformer
     /**
      * {@inheritdoc}
      */
-    public function extractDateOptions($matched, $length)
+    public function extractDateOptions(string $matched, int $length): array
     {
-        return array(
+        return [
             'hour' => (int) $matched,
             'hourInstance' => $this,
-        );
+        ];
     }
 }
