@@ -23,7 +23,7 @@ class SecondTransformer extends Transformer
     /**
      * {@inheritdoc}
      */
-    public function format(\DateTime $dateTime, $length)
+    public function format(\DateTime $dateTime, int $length): string
     {
         $secondOfMinute = (int) $dateTime->format('s');
 
@@ -33,7 +33,7 @@ class SecondTransformer extends Transformer
     /**
      * {@inheritdoc}
      */
-    public function getReverseMatchingRegExp($length)
+    public function getReverseMatchingRegExp(int $length): string
     {
         return 1 === $length ? '\d{1,2}' : '\d{'.$length.'}';
     }
@@ -41,10 +41,10 @@ class SecondTransformer extends Transformer
     /**
      * {@inheritdoc}
      */
-    public function extractDateOptions($matched, $length)
+    public function extractDateOptions(string $matched, int $length): array
     {
-        return array(
+        return [
             'second' => (int) $matched,
-        );
+        ];
     }
 }
