@@ -29,7 +29,7 @@ abstract class Transformer
      *
      * @return string The formatted value
      */
-    abstract public function format(\DateTime $dateTime, $length);
+    abstract public function format(\DateTime $dateTime, int $length): string;
 
     /**
      * Returns a reverse matching regular expression of a string generated by format().
@@ -38,7 +38,7 @@ abstract class Transformer
      *
      * @return string The reverse matching regular expression
      */
-    abstract public function getReverseMatchingRegExp($length);
+    abstract public function getReverseMatchingRegExp(int $length): string;
 
     /**
      * Extract date options from a matched value returned by the processing of the reverse matching
@@ -49,7 +49,7 @@ abstract class Transformer
      *
      * @return array An associative array
      */
-    abstract public function extractDateOptions($matched, $length);
+    abstract public function extractDateOptions(string $matched, int $length): array;
 
     /**
      * Pad a string with zeros to the left.
@@ -59,7 +59,7 @@ abstract class Transformer
      *
      * @return string The padded string
      */
-    protected function padLeft($value, $length)
+    protected function padLeft(string $value, int $length): string
     {
         return str_pad($value, $length, '0', STR_PAD_LEFT);
     }
