@@ -19,7 +19,7 @@ class MethodNotImplementedException extends NotImplementedException
     /**
      * @param string $methodName The name of the method
      */
-    public function __construct($methodName)
+    public function __construct(string $methodName)
     {
         parent::__construct(sprintf('The %s() is not implemented.', $methodName));
     }
