@@ -18,7 +18,7 @@ namespace Symfony\Component\Intl\Exception;
  */
 class UnexpectedTypeException extends InvalidArgumentException
 {
-    public function __construct($value, $expectedType)
+    public function __construct($value, string $expectedType)
     {
         parent::__construct(sprintf('Expected argument of type "%s", "%s" given', $expectedType, \is_object($value) ? \get_class($value) : \gettype($value)));
     }
