@@ -55,7 +55,7 @@ function run($command)
 
 function get_icu_version_from_genrb($genrb)
 {
-    exec($genrb.' --version 2>&1', $output, $status);
+    exec($genrb.' --version - 2>&1', $output, $status);
 
     if (0 !== $status) {
         bailout($genrb.' failed.');
@@ -67,6 +67,12 @@ function get_icu_version_from_genrb($genrb)
 
     return $matches[1];
 }
+
+error_reporting(E_ALL);
+
+set_error_handler(function ($type, $msg, $file, $line) {
+    throw new \ErrorException($msg, 0, $type, $file, $line);
+});
 
 set_exception_handler(function (\Throwable $exception) {
     echo "\n";
@@ -82,10 +88,7 @@ set_exception_handler(function (\Throwable $exception) {
         echo get_class($cause).': '.$cause->getMessage()."\n";
         echo "\n";
         echo $cause->getFile().':'.$cause->getLine()."\n";
-        foreach ($cause->getTrace() as $trace) {
-            echo $trace['file'].':'.$trace['line']."\n";
-        }
-        echo "\n";
+        echo $cause->getTraceAsString()."\n";
 
         $cause = $cause->getPrevious();
         $root = false;
