@@ -33,24 +33,24 @@ abstract class AbstractCollatorTest extends TestCase
 
     public function asortProvider()
     {
-        return array(
+        return [
             /* array, sortFlag, expected */
-            array(
-                array('a', 'b', 'c'),
+            [
+                ['a', 'b', 'c'],
                 Collator::SORT_REGULAR,
-                array('a', 'b', 'c'),
-            ),
-            array(
-                array('c', 'b', 'a'),
+                ['a', 'b', 'c'],
+            ],
+            [
+                ['c', 'b', 'a'],
                 Collator::SORT_REGULAR,
-                array(2 => 'a', 1 => 'b',  0 => 'c'),
-            ),
-            array(
-                array('b', 'c', 'a'),
+                [2 => 'a', 1 => 'b',  0 => 'c'],
+            ],
+            [
+                ['b', 'c', 'a'],
                 Collator::SORT_REGULAR,
-                array(2 => 'a', 0 => 'b', 1 => 'c'),
-            ),
-        );
+                [2 => 'a', 0 => 'b', 1 => 'c'],
+            ],
+        ];
     }
 
     /**
