@@ -1,17 +1,17 @@
 <?php
 
-return array(
-    'Entry1' => array(
-        'Array' => array(
+return [
+    'Entry1' => [
+        'Array' => [
             0 => 'foo',
             1 => 'bar',
-        ),
+        ],
         'Integer' => 5,
         'Boolean' => false,
         'Float' => 1.23,
-    ),
+    ],
     'Entry2' => 'String',
-    'Traversable' => array(
+    'Traversable' => [
         'Foo' => 'Bar',
-    ),
-);
+    ],
+];
