@@ -22,13 +22,13 @@ abstract class AbstractIntlGlobalsTest extends TestCase
 {
     public function errorNameProvider()
     {
-        return array(
-            array(-129, '[BOGUS UErrorCode]'),
-            array(0, 'U_ZERO_ERROR'),
-            array(1, 'U_ILLEGAL_ARGUMENT_ERROR'),
-            array(9, 'U_PARSE_ERROR'),
-            array(129, '[BOGUS UErrorCode]'),
-        );
+        return [
+            [-129, '[BOGUS UErrorCode]'],
+            [0, 'U_ZERO_ERROR'],
+            [1, 'U_ILLEGAL_ARGUMENT_ERROR'],
+            [9, 'U_PARSE_ERROR'],
+            [129, '[BOGUS UErrorCode]'],
+        ];
     }
 
     /**
