@@ -21,24 +21,24 @@ class IcuVersionTest extends TestCase
 {
     public function normalizeProvider()
     {
-        return array(
-            array(null, '1', '10'),
-            array(null, '1.2', '12'),
-            array(null, '1.2.3', '12.3'),
-            array(null, '1.2.3.4', '12.3.4'),
-            array(1, '1', '10'),
-            array(1, '1.2', '12'),
-            array(1, '1.2.3', '12'),
-            array(1, '1.2.3.4', '12'),
-            array(2, '1', '10'),
-            array(2, '1.2', '12'),
-            array(2, '1.2.3', '12.3'),
-            array(2, '1.2.3.4', '12.3'),
-            array(3, '1', '10'),
-            array(3, '1.2', '12'),
-            array(3, '1.2.3', '12.3'),
-            array(3, '1.2.3.4', '12.3.4'),
-        );
+        return [
+            [null, '1', '10'],
+            [null, '1.2', '12'],
+            [null, '1.2.3', '12.3'],
+            [null, '1.2.3.4', '12.3.4'],
+            [1, '1', '10'],
+            [1, '1.2', '12'],
+            [1, '1.2.3', '12'],
+            [1, '1.2.3.4', '12'],
+            [2, '1', '10'],
+            [2, '1.2', '12'],
+            [2, '1.2.3', '12.3'],
+            [2, '1.2.3.4', '12.3'],
+            [3, '1', '10'],
+            [3, '1.2', '12'],
+            [3, '1.2.3', '12.3'],
+            [3, '1.2.3.4', '12.3.4'],
+        ];
     }
 
     /**
@@ -51,55 +51,55 @@ class IcuVersionTest extends TestCase
 
     public function compareProvider()
     {
-        return array(
-            array(null, '1', '==', '1', true),
-            array(null, '1.0', '==', '1.1', false),
-            array(null, '1.0.0', '==', '1.0.1', false),
-            array(null, '1.0.0.0', '==', '1.0.0.1', false),
-            array(null, '1.0.0.0.0', '==', '1.0.0.0.1', false),
+        return [
+            [null, '1', '==', '1', true],
+            [null, '1.0', '==', '1.1', false],
+            [null, '1.0.0', '==', '1.0.1', false],
+            [null, '1.0.0.0', '==', '1.0.0.1', false],
+            [null, '1.0.0.0.0', '==', '1.0.0.0.1', false],
 
-            array(null, '1', '==', '10', true),
-            array(null, '1.0', '==', '11', false),
-            array(null, '1.0.0', '==', '10.1', false),
-            array(null, '1.0.0.0', '==', '10.0.1', false),
-            array(null, '1.0.0.0.0', '==', '10.0.0.1', false),
+            [null, '1', '==', '10', true],
+            [null, '1.0', '==', '11', false],
+            [null, '1.0.0', '==', '10.1', false],
+            [null, '1.0.0.0', '==', '10.0.1', false],
+            [null, '1.0.0.0.0', '==', '10.0.0.1', false],
 
-            array(1, '1', '==', '1', true),
-            array(1, '1.0', '==', '1.1', false),
-            array(1, '1.0.0', '==', '1.0.1', true),
-            array(1, '1.0.0.0', '==', '1.0.0.1', true),
-            array(1, '1.0.0.0.0', '==', '1.0.0.0.1', true),
+            [1, '1', '==', '1', true],
+            [1, '1.0', '==', '1.1', false],
+            [1, '1.0.0', '==', '1.0.1', true],
+            [1, '1.0.0.0', '==', '1.0.0.1', true],
+            [1, '1.0.0.0.0', '==', '1.0.0.0.1', true],
 
-            array(1, '1', '==', '10', true),
-            array(1, '1.0', '==', '11', false),
-            array(1, '1.0.0', '==', '10.1', true),
-            array(1, '1.0.0.0', '==', '10.0.1', true),
-            array(1, '1.0.0.0.0', '==', '10.0.0.1', true),
+            [1, '1', '==', '10', true],
+            [1, '1.0', '==', '11', false],
+            [1, '1.0.0', '==', '10.1', true],
+            [1, '1.0.0.0', '==', '10.0.1', true],
+            [1, '1.0.0.0.0', '==', '10.0.0.1', true],
 
-            array(2, '1', '==', '1', true),
-            array(2, '1.0', '==', '1.1', false),
-            array(2, '1.0.0', '==', '1.0.1', false),
-            array(2, '1.0.0.0', '==', '1.0.0.1', true),
-            array(2, '1.0.0.0.0', '==', '1.0.0.0.1', true),
+            [2, '1', '==', '1', true],
+            [2, '1.0', '==', '1.1', false],
+            [2, '1.0.0', '==', '1.0.1', false],
+            [2, '1.0.0.0', '==', '1.0.0.1', true],
+            [2, '1.0.0.0.0', '==', '1.0.0.0.1', true],
 
-            array(2, '1', '==', '10', true),
-            array(2, '1.0', '==', '11', false),
-            array(2, '1.0.0', '==', '10.1', false),
-            array(2, '1.0.0.0', '==', '10.0.1', true),
-            array(2, '1.0.0.0.0', '==', '10.0.0.1', true),
+            [2, '1', '==', '10', true],
+            [2, '1.0', '==', '11', false],
+            [2, '1.0.0', '==', '10.1', false],
+            [2, '1.0.0.0', '==', '10.0.1', true],
+            [2, '1.0.0.0.0', '==', '10.0.0.1', true],
 
-            array(3, '1', '==', '1', true),
-            array(3, '1.0', '==', '1.1', false),
-            array(3, '1.0.0', '==', '1.0.1', false),
-            array(3, '1.0.0.0', '==', '1.0.0.1', false),
-            array(3, '1.0.0.0.0', '==', '1.0.0.0.1', true),
+            [3, '1', '==', '1', true],
+            [3, '1.0', '==', '1.1', false],
+            [3, '1.0.0', '==', '1.0.1', false],
+            [3, '1.0.0.0', '==', '1.0.0.1', false],
+            [3, '1.0.0.0.0', '==', '1.0.0.0.1', true],
 
-            array(3, '1', '==', '10', true),
-            array(3, '1.0', '==', '11', false),
-            array(3, '1.0.0', '==', '10.1', false),
-            array(3, '1.0.0.0', '==', '10.0.1', false),
-            array(3, '1.0.0.0.0', '==', '10.0.0.1', true),
-        );
+            [3, '1', '==', '10', true],
+            [3, '1.0', '==', '11', false],
+            [3, '1.0.0', '==', '10.1', false],
+            [3, '1.0.0.0', '==', '10.0.1', false],
+            [3, '1.0.0.0.0', '==', '10.0.0.1', true],
+        ];
     }
 
     /**
