@@ -21,28 +21,28 @@ class VersionTest extends TestCase
 {
     public function normalizeProvider()
     {
-        return array(
-            array(null, '1', '1'),
-            array(null, '1.2', '1.2'),
-            array(null, '1.2.3', '1.2.3'),
-            array(null, '1.2.3.4', '1.2.3.4'),
-            array(1, '1', '1'),
-            array(1, '1.2', '1'),
-            array(1, '1.2.3', '1'),
-            array(1, '1.2.3.4', '1'),
-            array(2, '1', '1'),
-            array(2, '1.2', '1.2'),
-            array(2, '1.2.3', '1.2'),
-            array(2, '1.2.3.4', '1.2'),
-            array(3, '1', '1'),
-            array(3, '1.2', '1.2'),
-            array(3, '1.2.3', '1.2.3'),
-            array(3, '1.2.3.4', '1.2.3'),
-            array(4, '1', '1'),
-            array(4, '1.2', '1.2'),
-            array(4, '1.2.3', '1.2.3'),
-            array(4, '1.2.3.4', '1.2.3.4'),
-        );
+        return [
+            [null, '1', '1'],
+            [null, '1.2', '1.2'],
+            [null, '1.2.3', '1.2.3'],
+            [null, '1.2.3.4', '1.2.3.4'],
+            [1, '1', '1'],
+            [1, '1.2', '1'],
+            [1, '1.2.3', '1'],
+            [1, '1.2.3.4', '1'],
+            [2, '1', '1'],
+            [2, '1.2', '1.2'],
+            [2, '1.2.3', '1.2'],
+            [2, '1.2.3.4', '1.2'],
+            [3, '1', '1'],
+            [3, '1.2', '1.2'],
+            [3, '1.2.3', '1.2.3'],
+            [3, '1.2.3.4', '1.2.3'],
+            [4, '1', '1'],
+            [4, '1.2', '1.2'],
+            [4, '1.2.3', '1.2.3'],
+            [4, '1.2.3.4', '1.2.3.4'],
+        ];
     }
 
     /**
@@ -55,27 +55,27 @@ class VersionTest extends TestCase
 
     public function compareProvider()
     {
-        return array(
-            array(null, '1', '==', '1', true),
-            array(null, '1.0', '==', '1.1', false),
-            array(null, '1.0.0', '==', '1.0.1', false),
-            array(null, '1.0.0.0', '==', '1.0.0.1', false),
+        return [
+            [null, '1', '==', '1', true],
+            [null, '1.0', '==', '1.1', false],
+            [null, '1.0.0', '==', '1.0.1', false],
+            [null, '1.0.0.0', '==', '1.0.0.1', false],
 
-            array(1, '1', '==', '1', true),
-            array(1, '1.0', '==', '1.1', true),
-            array(1, '1.0.0', '==', '1.0.1', true),
-            array(1, '1.0.0.0', '==', '1.0.0.1', true),
+            [1, '1', '==', '1', true],
+            [1, '1.0', '==', '1.1', true],
+            [1, '1.0.0', '==', '1.0.1', true],
+            [1, '1.0.0.0', '==', '1.0.0.1', true],
 
-            array(2, '1', '==', '1', true),
-            array(2, '1.0', '==', '1.1', false),
-            array(2, '1.0.0', '==', '1.0.1', true),
-            array(2, '1.0.0.0', '==', '1.0.0.1', true),
+            [2, '1', '==', '1', true],
+            [2, '1.0', '==', '1.1', false],
+            [2, '1.0.0', '==', '1.0.1', true],
+            [2, '1.0.0.0', '==', '1.0.0.1', true],
 
-            array(3, '1', '==', '1', true),
-            array(3, '1.0', '==', '1.1', false),
-            array(3, '1.0.0', '==', '1.0.1', false),
-            array(3, '1.0.0.0', '==', '1.0.0.1', true),
-        );
+            [3, '1', '==', '1', true],
+            [3, '1.0', '==', '1.1', false],
+            [3, '1.0.0', '==', '1.0.1', false],
+            [3, '1.0.0.0', '==', '1.0.0.1', true],
+        ];
     }
 
     /**
